Disable DocReader methods when disposed (#47)
Before this change it was possible to use the DocReader methods when
the reader was already disposed. This caused confusing behavior like
PageCount being 0 even when it should be something else and there
was no error or exception indicating that something was wrong.

Now using the methods has been prevented by checking for disposed
flag and throwing exception in that case.